fix. Code cleaning
