+++ src/book_summarizer/tools/summarizer/summary_tree.py
@@ -3,7 +3,7 @@ import numpy as np
 import ruptures as rpt
 import networkx as nx
 from plotly import graph_objects as go
-from sklearn.metrics.pairwise import cosine_similarity
+
 from tqdm import tqdm
 
 
@@ -23,44 +23,6 @@ def get_segments_from_breakpoints(bkpts: np.array):
     return segments
 
 
-def medoid(arr: np.array):
-    """
-    Calculate the medoid of a set of vectors.
-
-    The medoid is the point in the set with the smallest average distance to all other points.
-
-    Args:
-        arr (np.array): A 2D array of vectors.
-
-    Returns:
-        np.array: The medoid vector.
-    """
-    dist_matrix = cosine_similarity(arr)
-    idx_medoid = np.argmin(np.sum(dist_matrix, axis=1))
-    return arr[idx_medoid]
-
-
-def compute_centered_signal(signal: np.array, bkpts: np.array):
-    """
-    Center a signal based on the medoid of each segment defined by the breakpoints.
-
-    Args:
-        signal (np.array): The input signal.
-        bkpts (np.array): An array of breakpoint indices.
-
-    Returns:
-        np.array: The centered signal.
-    """
-    segments = get_segments_from_breakpoints(bkpts)
-    centered_signal = np.zeros_like(signal)
-
-    for seg in segments:
-        _signal = signal[seg[0] : seg[1] + 1]
-        center = medoid(_signal)
-        centered_signal[seg[0] : seg[1] + 1] = center
-    return centered_signal
-
-
 class SummaryTree:
     """
     A class to represent the hierarchical structure of a text using change point detection.
@@ -72,38 +34,16 @@ class SummaryTree:
 
     def __init__(
         self,
-        penalties: np.array = np.arange(2, 50, 0.25),
-        denoise: bool = True,
-        algorithm: str = "binseg",
-        metric: str = "cosine",
+        bkpts_matrix: np.array,
+        bkpts: list,
+        penalties: np.array,
     ):
-        """
-        Initialize the SummaryTree.
-
-        Args:
-            penalties (np.array, optional): An array of penalty values to use for change point detection.
-                Defaults to np.arange(2, 50, 0.25).
-            denoise (bool, optional): Whether to denoise the signal before change point detection.
-                Defaults to True.
-            algorithm (str, optional): The algorithm to use for change point detection. Defaults to "binseg".
-            metric (str, optional): The metric to use for computing the cost of change points. Defaults to "cosine".
-        """
+        self.bkpts_matrix = bkpts_matrix
+        self.bkpts = bkpts
         self.penalties = penalties
-        self.denoise = denoise
-        self.algorithm = algorithm
-        self.metric = metric
+        self._build_tree()
 
-    def fit(self, signal: np.array):
-        """
-        Fit the SummaryTree to the given signal.
-
-        Args:
-            signal (np.array): The input signal (e.g., embeddings of text chunks).
-
-        Returns:
-            SummaryTree: The fitted SummaryTree object.
-        """
-        self._find_change_points(signal)
+    def fit(self):
         self.graph = self._build_tree()
         return self
 
@@ -245,44 +185,6 @@ class SummaryTree:
             node_levels = {node: len(shortest_path[node]) - 1 for node in shortest_path}
             nx.set_node_attributes(self.graph, node_levels, "level")
 
-    def _find_change_points(self, signal: np.array) -> None:
-        """
-        Find change points in the given signal using the specified penalty values.
-
-        This method iterates through the penalty values, performing change point detection
-        at each level and optionally denoising the signal between levels.
-
-        Args:
-            signal (np.array): The input signal.
-        """
-        bkps = []
-        centered_signal = signal
-
-        for pen in tqdm(self.penalties, desc="Finding Change Points"):
-            # cost = rpt.costs.CostCosine()
-            if self.algorithm == "binseg":
-                model = rpt.Binseg(model=self.metric, min_size=2, jump=1)
-            elif self.algorithm == "window":
-                model = rpt.Window(width=20, model=self.metric)
-            elif self.algorithm == "pelt":
-                model = rpt.KernelCPD(kernel=self.metric, min_size=2)
-            else:
-                raise ValueError("Invalid algorithm. Must be 'binseg' or 'pelt'.")
-
-            model.fit(centered_signal)
-            bkps.append(model.predict(pen=pen))
-
-            if self.denoise:
-                centered_signal = compute_centered_signal(centered_signal, bkps[-1])
-
-        bkpt_matrix = np.zeros((len(self.penalties), len(signal)))
-        for i, _bkps in enumerate(bkps):
-            for j in _bkps[:-1]:
-                bkpt_matrix[i, j - 1] = 1
-
-        self.bkpts = bkps
-        self.bkpts_matrix = bkpt_matrix
-
     def _build_tree(self) -> None:
         """
         Build the summary tree from the detected change points.
@@ -345,12 +247,12 @@ class SummaryTree:
                         G.add_edge(
                             segment,
                             prev_segment,
-                            label="summarized_from",
+                            label="parent_section",
                             weight=(min(end, prev_end) - max(start, prev_start))
                             / (end - start),
                         )
 
-        # Add order atrtibute to the last level
+        # Add order attribute to the last level
         for order, segment in enumerate(current_segments):
             nx.set_node_attributes(G, {segment: {"order": order}})
